rename : 패키지 생성 baseballGames class -> impl 이동
